+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Project versions conform to [Semantic Versioning](https://semver.org/)
 
 - `Removed`: for deprecated features removed in this release
 
+## [1.14.2]
+
+- added a function for managing fields with parts in BasicHelpers `normalizeFieldWithParts`
+- added a function to get a field out of a production `getFieldPartSubfield` which can be used for similar top-level fields (like creation)
+
 ## [0.14.1]
 
 - adds `ObjectHelpers.js` with initial `getDimensionDescriptions` helper and tests